completed ex-5.11 to 5.16

+++ part 5-bloglist-frontend/src/components/Blog.js
@@ -1,34 +1,37 @@
-import React,{useState} from 'react'
+import React,{ useState } from 'react'
 
 const blogStyle={
-    paddingTop:10,
-    paddngLeft:2,
-    border:'solid',
-    borderWidth:1,
-    marginBottom:5
-  }
+  paddingTop:10,
+  paddngLeft:2,
+  border:'solid',
+  borderWidth:1,
+  marginBottom:5
+}
 
-const BlogBeforeToggle = ({blog,toggleVisibility,text}) => {
+const BlogBeforeToggle = ({ blog,toggleVisibility,text }) => {
 
   return (
-  <div style={blogStyle}>
-    {blog.title} - {blog.author} <button onClick={()=>toggleVisibility()}>{text}</button> <br/>
-  </div>  
+    <div style={blogStyle} className="before">
+      {blog.title} - {blog.author} <button onClick={() => toggleVisibility()}>{text}</button> <br/>
+    </div>
   )
 }
 
-const BlogAfterToggle = ({blog,toggleVisibility,text,updateBlog,deleteBlog}) => {
+const BlogAfterToggle = ({ blog,toggleVisibility,text,updateBlog,deleteBlog }) => {
   const [likes,setLikes]=useState(blog.likes)
   const [del,setDel]=useState(false)
-  const handleLikes=()=>{
-    blog.likes=likes+1;
+  const handleLikes=() => {
+    blog.likes=likes+1
     updateBlog(blog.id,blog)
     setLikes(blog.likes)
   }
 
-  const handleDelete=()=>{
-    deleteBlog(blog.id)
-    setDel(true)
+  const handleDelete=() => {
+    if(window.confirm(`Remove blog ${blog.title}`))
+    {
+      deleteBlog(blog.id)
+      setDel(true)
+    }
   }
 
   if(del)
@@ -36,28 +39,30 @@ const BlogAfterToggle = ({blog,toggleVisibility,text,updateBlog,deleteBlog}) => 
     return null
   }
   else {
-  return (
-      <div style={blogStyle}>
-        {blog.title} - {blog.author} <button onClick={()=>toggleVisibility()}>{text}</button> <br/>
+    return (
+      <div style={blogStyle} className="after">
+        {blog.title} - {blog.author} <button onClick={() => toggleVisibility()}>{text}</button> <br/>
         {blog.url} <br/>
-        likes {blog.likes}  <button onClick={()=>handleLikes(updateBlog)}>like</button> <br/>
+        likes {blog.likes}  <button onClick={() => handleLikes(updateBlog)}>like</button> <br/>
         {blog.author} <br/>
         <button onClick={handleDelete}>remove</button>
-      </div>  
-  )
+      </div>
+    )
   }
 }
 
-const Blog=({blog,updateBlog,deleteBlog})=>{
+const Blog=({ blog,updateBlog,deleteBlog }) => {
   const [toggle,setToggle]=useState(false)
   const text=toggle?'hide':'view'
-  const toggleVisibility=()=>{
+  const toggleVisibility=() => {
     setToggle(!toggle)
   }
   return (
-        toggle
-        ?<BlogAfterToggle blog={blog} toggleVisibility={toggleVisibility} text={text} updateBlog={updateBlog} deleteBlog={deleteBlog}/>
-        :<BlogBeforeToggle blog={blog} toggleVisibility={toggleVisibility} text={text} />
+    <div className="BlogComponent">
+    toggle
+      ?<BlogAfterToggle blog={blog} toggleVisibility={toggleVisibility} text={text} updateBlog={updateBlog} deleteBlog={deleteBlog}/>
+      :<BlogBeforeToggle blog={blog} toggleVisibility={toggleVisibility} text={text} />
+    </div>
   )
 }
